requirements.txt
diffusers
torch
transformers
gradio
Pillow
python-dotenv